change Bounds args to int instead of int32

+++ orphan.go
@@ -26,7 +26,7 @@ func Clear(r, g, b, a float32) {
 // Bounds sets the drawing bounds in pixels. Drawing outside bounds is always discarted.
 //
 // Calling this function is equivalent to setting viewport and scissor in OpenGL.
-func Bounds(x, y, w, h int32) {
-	gl.Viewport(x, y, w, h)
-	gl.Scissor(x, y, w, h)
+func Bounds(x, y, w, h int) {
+	gl.Viewport(int32(x), int32(y), int32(w), int32(h))
+	gl.Scissor(int32(x), int32(y), int32(w), int32(h))
 }
